flatten: touchRipple
