+++ packages/client/lib/app/core/widgets/touch_ripple/canvas/touch_ripple_painter.dart
