--- packages/client/lib/app/core/widgets/touch_ripple/stack/presentation/presentation.dart
@@ -1,2 +1,0 @@
-export 'mobx/touch_ripple_store.dart';
-export 'widget/touch_ripple.dart';
