--- packages/client/lib/app/core/widgets/touch_ripple/stack/stack.dart
@@ -1,4 +1,0 @@
-export 'constants/constants.dart';
-export './movies/touch_ripple_movie.dart';
-export 'presentation/presentation.dart';
-export './types/types.dart';
