+++ packages/client/lib/app/core/widgets/touch_ripple/types/touch_ripple_animation_info.dart
